Push day1 to repo
I was also playing around with some stuff in rust, such as have a method
take the input and a method to resolve the input to 'bootstrap' each
part of the day. Didn't get it working though.

I also really need to play catch-up, as always, fell behind the days!